Messages

+++ BITC-Messenger-IOS/FileDataStore.swift
@@ -18,6 +18,7 @@ public class LocalMessageStore{
         var from    : String = ""
         var to      : String = ""
         var N       : Int    = 0
+        
         init (prefix:String, id:String, from:String, to:String, N:Int){
             self.prefix = prefix
             self.id     = id
@@ -40,6 +41,9 @@ public class LocalMessageStore{
         }
         func saveFile(data:Data){
             _ = VSMAPI.saveFile(name:"\(prefix)_\(id)_\(from)_\(to)_\(N)_.I", data: data)
+        }
+        func getFileData()->Data?{
+            return VSMAPI.getFile(name: "\(prefix)_\(id)_\(from)_\(to)_\(N)_.I")
         }
     
     }
@@ -67,9 +71,6 @@ public class LocalMessageStore{
         }
         db = tmparr.sorted(by:{$0.from < $1.from || $0.from == ""})
     }
-    private func callback(){
-        getData(isAfter:isAfter, keyMsg: keyMsg, jamp : jamp)
-    }
     private func loadPart(){
         if VSMAPI.Connectivity.isConn{
             let prms = ["ConversationId":id, "N":chank, "IsAfter":self.isAfter ? "True" : "False", "MessageId":self.keyMsg, "Email" : VSMAPI.Settings.user, "PasswordHash" : VSMAPI.Settings.hash] as [String : Any]
@@ -84,28 +85,34 @@ public class LocalMessageStore{
                         let data = d as! Data
                         if let json = try? JSON(data: data) {
                             let arr = json["Messages"].array!
-                            if let part = self.db.first(where:{$0.from == ""}){
-                                part.deleteFile()
-                                self.db = self.db.filter { $0 !== part }
-                            }
                             let _N      = arr.count
                             let _from   = arr.first?.dictionary?["Id"]?.string ?? ""
                             let _to     = arr.last?.dictionary?["Id"]?.string ?? ""
-                            let newPart = Part(prefix: self.table, id: self.id, from: _from, to: _to, N: _N)
-                            newPart.saveFile(data: data)
-                            if(self.isAfter){
-                                self.db.append(newPart)
+                            
+                            if _N > 0{
+                                
+                                let newPart = Part(prefix: self.table, id: self.id, from: _from, to: _to, N: _N)
+                                
+                                if let part = self.db.first(where:{$0.from == ""}){
+                                    part.deleteFile()
+                                    self.db = self.db.filter { $0 !== part }
+                                }
+                                newPart.saveFile(data: data)
+                            
+                                if(self.isAfter){
+                                    self.db.append(newPart)
+                                }
+                                else{
+                                    self.db.insert(newPart, at: 0)
+                                }
+                                if (self.allNext || self.isAfter) && _N > 0{
+                                    self.loadPart()
+                                    return
+                                }
+                                else {
+                                    self.getData(ext:false, isAfter:self.isAfter, keyMsg: self.isAfter ? _to : _from, jamp : self.jamp)
+                                }
                             }
-                            else{
-                                self.db.insert(newPart, at: 0)
-                            }
-                            if (self.allNext || self.isAfter) && _N > 0{
-                                self.loadPart()
-                            }
-                            else{
-                                self.callback()
-                            }
-                            return
                         }
                     }
                 }
@@ -113,6 +120,33 @@ public class LocalMessageStore{
             }()}
             )
         }
+    }
+    private func filArrayFromFile(_ p:Part){
+        if let data = p.getFileData(){
+            if let json = try? JSON(data: data) {
+                let arr = json["Messages"].array!
+                for c in arr{
+                    if let dict = c.dictionary{
+                        Messages.append(VSMMessage(from:dict))
+                    }
+                }
+            }
+        }
+        
+    }
+    private func fillArray(_ i:Int){
+        let wt = db[i]
+        let wu = i == 0             ?   nil : db[i+1]
+        let wd = i == db.count-1    ?   nil : db[i-1]
+        
+        self.Messages.removeAll()
+        
+        /*filArrayFromFile(wu)
+        filArrayFromFile(wt)
+        filArrayFromFile(wd)*/
+        
+        let addr = self.jamp ? Messages.count - 1 : isAfter ? Messages.count - (wd?.N ?? 0) : 0 + (wu?.N ?? 0)
+        delegate(Messages, addr)
     }
     public var Messages = [VSMMessage]()
     public init(Id:String, Table:String = "MSG", Chank:Int = 20, Separator:String = "_", Delegate:@escaping ([VSMMessage], Int) ->()){
@@ -123,23 +157,51 @@ public class LocalMessageStore{
         delegate    = Delegate
         fillDB()
     }
-    public func getData(isAfter:Bool = false, keyMsg: String = "", jamp : Bool = false){
-        self.isAfter = isAfter
-        self.keyMsg  = keyMsg
-        self.jamp    = jamp
-        self.allNext = false
-        if keyMsg == ""{
-            if db.count == 0{
+    public func getData(ext:Bool = true, isAfter:Bool = false, keyMsg: String = "", jamp : Bool = false){
+        if ext{
+            self.isAfter = isAfter
+            self.keyMsg  = keyMsg
+            self.jamp    = jamp
+            self.allNext = false //true для поиска.
+
+            if db.count > 0 {
+                if let part = (self.keyMsg == "" || self.jamp) ? db.last : db.first(where: {(isAfter ? $0.to : $0.from) == keyMsg}){
+                if let i = db.index(where: {(isAfter ? $0.to == part.to : $0.from == part.from)}) {
+                    if part.N < self.chank && db.count > 1{// блок не заполнен и не единственный надо подкачать данные с сервера
+                        let fullPart = db[isAfter ? i - 1 : i + 1]
+                        self.keyMsg = isAfter ? fullPart.to : fullPart.from
+                        loadPart()
+                    }
+                    else if part.N < self.chank && db.count == 1{ //единственный блок не заполнен. надо подкачать данные с сервера
+                        self.keyMsg = ""
+                        loadPart()
+                    }
+                    else{//блок заполнен ищем сверху/снизу
+                        if i == 0 || i == db.count - 1 {// это самые старые сообщения или самые новые
+                            loadPart()
+                        }
+                        else{//берем из сохраненных
+                            fillArray(isAfter ? i - 1 : i + 1)
+                        }
+                        
+                    }
+                }
+            }
+            }
+            else{//нет ничего пробуем получить что нибудь
+                self.keyMsg = ""
                 loadPart()
             }
-            //заполняем массив из единственного файла и отдаем его
         }
-        else if let part = db.first(where: {(isAfter ? $0.to : $0.from) == keyMsg}){
-            let i = db.index(where: {(isAfter ? $0.to : $0.from) == keyMsg})
-            //if
-        }
-        else{//нет ничего
-            loadPart()
+        else{
+            if let part = (self.keyMsg == "" || self.jamp) ? db.last : db.first(where: {(isAfter ? $0.to : $0.from) == keyMsg}){
+                if let i = db.index(where: {(isAfter ? $0.to == part.to : $0.from == part.from)}) {
+                    fillArray(i)
+                }
+            }
+            else{
+                delegate(self.Messages, isAfter ? Messages.count - 1 : 0)
+            }
         }
     }
     public func getFilteredData(filter:String, isAfter:Bool = false, keyMsg: String = "", delegate:([VSMMessage], Int) ->()){
